requirements.txt
asgiref==3.10.0
certifi==2025.10.5
charset-normalizer==3.4.4
dj-database-url==3.0.1
Django==5.2.7
django-cors-headers==4.9.0
idna==3.11
python-decouple==3.8
requests==2.32.5
sqlparse==0.5.3
urllib3==2.5.0
